pyproject.toml
[project]
name = "fastapi-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "black>=25.12.0",
    "celery>=5.6.2",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.124.4",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.15.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
env_files = [".env-test"]

[tool.ruff]
line-length = 100